+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(SDMetrics)
+
+test_check("SDMetrics")