+++ NEWS.md
@@ -2,4 +2,5 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Specified CC0 license.
+* New areaOfExpectedOccurrence function added.